move confirmation dialog
